Save and open emails

+++ SaintSender.Core/Entities/Mail.cs
@@ -1,20 +1,17 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.IO;
-using System.Linq;
 using System.Runtime.Serialization;
 using System.Runtime.Serialization.Formatters.Binary;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace SaintSender.Core.Entities
 {
     [Serializable]
-    public class Mail : IDeserializationCallback, ISerializable
+    public class Mail :  ISerializable
     {
 
-        private static string path = @"C:\Users\Máté\Desktop\Advance\3_TW\1.txt";
+        private static string pathPart = "./data/";
+
 
         public int ID
         {
@@ -41,9 +38,8 @@ namespace SaintSender.Core.Entities
             get;set;
         }
 
-        public Mail() { }
 
-      
+        public Mail() { }
 
         public Mail(SerializationInfo info, StreamingContext context)
         {
@@ -55,6 +51,7 @@ namespace SaintSender.Core.Entities
             
         }
 
+
         public void GetObjectData(SerializationInfo info, StreamingContext context)
         {
             info.AddValue("id", this.ID);
@@ -64,37 +61,24 @@ namespace SaintSender.Core.Entities
             info.AddValue("body", this.Body);
         }
 
-        public static void Serialize(string output, ObservableCollection<Mail> mails )
+        public static void Serialize(string output, ObservableCollection<Mail> mails, String email)
         {
-
+            String path = pathPart + email + ".txt";
             IFormatter formatter = new BinaryFormatter();
-            if (!File.Exists(path))
-            {
-                Stream stream = new FileStream(path, FileMode.Create, FileAccess.Write);
-                formatter.Serialize(stream, mails);
-                stream.Close();
-            }
-            else
-            {
-                File.Delete(path);
-                Stream stream = new FileStream(path, FileMode.Create, FileAccess.Write);
-                formatter.Serialize(stream, mails);
-                stream.Close();
-            }
+            if (File.Exists(path)) File.Delete(path);
+            Stream stream = new FileStream(path, FileMode.Create, FileAccess.Write);
+            formatter.Serialize(stream, mails);
+            stream.Close();
         }
 
-        public static ObservableCollection<Mail> Deserialize()
+        public static ObservableCollection<Mail> Deserialize(String email)
         {
+            String path = pathPart + email + ".txt";
             ObservableCollection<Mail> _emailsToShow = null;
-
-            // Open the file containing the data that you want to deserialize.
             FileStream fs = new FileStream(path, FileMode.Open);
             try
             {
                 BinaryFormatter formatter = new BinaryFormatter();
-
-                // Deserialize the hashtable from the file and
-                // assign the reference to the local variable.
                 _emailsToShow = (ObservableCollection<Mail>)formatter.Deserialize(fs);
             }
             catch (SerializationException e)
@@ -102,18 +86,8 @@ namespace SaintSender.Core.Entities
                 Console.WriteLine("Failed to deserialize. Reason: " + e.Message);
                 throw;
             }
-            finally
-            {
-                fs.Close();
-            }
+            finally { fs.Close(); }
             return _emailsToShow;
-            
-           
-        }
-
-        void IDeserializationCallback.OnDeserialization(Object sender)
-        {
-            
         }
     }
 }
